requirements.txt
requests==2.32.3
pydantic==2.10.6
fastapi
python-dotenv==1.0.1
langchain==0.3.18
langchain-openai==0.3.4
langchain-community===0.3.17
langchain-huggingface==0.2.0
langchain-pinecone==0.2.2
pandas==2.2.3
pillow==11.1.0
pinecone==5.4.2
sentence-transformers==3.4.1
alembic
sqlalchemy
psycopg2-binary
pymysql
passlib
bcrypt
PyJWT
python-jose
pydantic[email]
uvicorn[standard]
